requirements.txt
Django==3.2.9
django-annoying==0.10.6
django-crispy-forms==1.13.0
django-extensions==3.1.5
djangorestframework==3.12.4
factory-boy==3.2.1
Faker==9.9.0
Pillow==8.4.0
psycopg2-binary==2.9.1
python-dotenv==0.19.1
gunicorn==20.1.0
pre_commit==2.15.0
django-heroku==0.3.1
weasyprint==54.0
redis==4.1.3
celery==5.0.5
django-celery-beat==2.2.0
django-storages==1.12.3
boto3==1.21.2
pytest-django==4.5.2
beautifulsoup4==4.10.0
django-jazzmin==2.5.0
postmarker==1.0
djangorestframework-simplejwt==5.2.2